pyproject.toml
[project]
name = "draco"
version = "1.0.0"
description = "Automated plant watering system for Raspberry Pi"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Joel Pérez Izquierdo", email = "joelperez91@gmail.com" }]
dependencies = [
    "RPi.GPIO>=0.7.1",
    "dbus-python>=1.3.2; sys_platform == 'linux'",
    "keyring>=24.2.0",
    "python-telegram-bot[job-queue]>=20.4",
    "schedule>=1.2.0",
    "paho-mqtt>=1.6.1",
    "pyyaml>=6.0",
]

[project.scripts]
draco = "draco:main"

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
